fixup clang warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/10376


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@34385 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/layout/LETypes.h
@@ -302,7 +302,7 @@ typedef struct LEPoint LEPoint;
 /**
  * Range check for overflow
  */
-#define LE_RANGE_CHECK(type, count, ptrfn) (( (LE_UINTPTR_MAX / sizeof(type)) < count ) ? NULL : (ptrfn))
+#define LE_RANGE_CHECK(type, count, ptrfn) (( (LE_UINTPTR_MAX / sizeof(type)) < (size_t)count ) ? NULL : (ptrfn))
 /**
  * A convenience macro to get the length of an array.
  *
